% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_ras.R
\name{aggregate_ras}
\alias{aggregate_ras}
\title{Reproject raster file}
\usage{
aggregate_ras(input_file, output_file, crs, res, ext, method)
}
\arguments{
\item{input_file}{Raster file, including path, to reproject}

\item{output_file}{Raster file, including path, to save as}

\item{crs}{Target reprojection}

\item{res}{Resolution value}

\item{ext}{Vector of four values: xmin, xmax, ymin, ymax}

\item{method}{Resampling method. "near" for categorical data, "bilinear" for continuous data. Other methods are possible, see \url{https://www.gdal.org/gdalwarp.html}}
}
\value{
Saved output file
}
\description{
Reproject raster file
}
\examples{
\dontrun{
aggregate_ras('my_path/temp.tif', 'my_path/reproj_temp.tif',
          crs = '+proj=utm +zone=55 +south +ellps=GRS80 +units=m +no_defs',
          res = 1000,
          method = "average",
          ext = c(-58000, 764000, 5661000, 6224000))
}
}
