% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_shp.R
\name{crop_shp}
\alias{crop_shp}
\title{Crop shapefile}
\usage{
crop_shp(input_file, output_file, ext)
}
\arguments{
\item{input_file}{Shapefile, including path, to reproject}

\item{output_file}{Shapefile, including path, to save as}

\item{ext}{Target extent (xmin, xmax, ymin, ymax) to which shapefile should be cropped. Same projection as input.}
}
\value{
Saved output file
}
\description{
Crop shapefile
}
\examples{
\dontrun{
crop_shp('my_path/lakes.shp', 'my_path/crop_lakes.shp',
          ext = c(5, 15, 47, 55)
}
}
