% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proximity_ras.R
\name{proximity_ras}
\alias{proximity_ras}
\title{Calculate proximity of raster file}
\usage{
proximity_ras(input_file, output_file)
}
\arguments{
\item{input_file}{Raster file, including path, to calculate from}

\item{output_file}{Raster file, including path, to save as}
}
\value{
Saved output file
}
\description{
Calculate proximity of raster file
}
\examples{
\dontrun{
proximity_ras('my_path/roads.tif', 'my_path/dist_roads.tif')
}
}
