% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_shp.R
\name{rasterize_shp}
\alias{rasterize_shp}
\title{Convert shapefile to raster}
\usage{
rasterize_shp(input_file, output_file, res, ext, attribute = NULL)
}
\arguments{
\item{input_file}{Shapefile file, including path, to convert to raster}

\item{output_file}{Raster file, including path, to save as}

\item{res}{Resolution value}

\item{ext}{Vector of four values: xmin, xmax, ymin, ymax}

\item{attribute}{by which to rasterize. Defaults to null}
}
\value{
Saved output file
}
\description{
Convert shapefile to raster
}
\examples{
\dontrun{
rasterize_shp('my_path/lakes.shp', 'my_path/lakes.tif',
             res = 1000,
             ext = c(-58000, 764000, 5661000, 6224000),
             attribute = "size")
}
}
