% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj_ras.R
\name{reproj_ras}
\alias{reproj_ras}
\title{Reproject raster file}
\usage{
reproj_ras(input_file, output_file, crs, res, ext)
}
\arguments{
\item{input_file}{Raster file, including path, to reproject}

\item{output_file}{Raster file, including path, to save as}

\item{crs}{Target reprojection}

\item{res}{Resolution value}

\item{ext}{Vector of four values: xmin, xmax, ymin, ymax}
}
\value{
Saved output file
}
\description{
Reproject raster file
}
\examples{
\dontrun{
reproj_shp('my_path/temp.tif', 'my_path/reproj_temp.tif',
          crs = CRS('+proj=utm +zone=55 +south +ellps=GRS80 +units=m +no_defs'),
          res = 1000,
          ext = c(-58000, 764000, 5661000, 6224000))
}
}
