% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproj_shp.R
\name{reproj_shp}
\alias{reproj_shp}
\title{Reproject shapefile}
\usage{
reproj_shp(input_file, output_file, crs)
}
\arguments{
\item{input_file}{Shapefile, including path, to reproject}

\item{output_file}{Shapefile, including path, to save as}

\item{crs}{Target reprojection}
}
\value{
Saved output file
}
\description{
Reproject shapefile
}
\examples{
\dontrun{
reproj_shp('my_path/lakes.shp', 'my_path/reproj_lakes.shp',
          crs = '+proj=utm +zone=55 +south +ellps=GRS80 +units=m +no_defs')
}
}
