% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_ras.R
\name{slope_ras}
\alias{slope_ras}
\title{Calculate slope}
\usage{
slope_ras(input_file, output_file)
}
\arguments{
\item{input_file}{Raster file, including path, to calculate slope from}

\item{output_file}{Raster file, including path, to save to}
}
\value{
Saved output file
}
\description{
Calculate slope
}
\examples{
\dontrun{
slope_ras('my_path/dem.img', 'my_path/slope.tif')
}
}
